changed magnetomter readings to ints

+++ pcm/tx.c
@@ -910,9 +910,9 @@ static void StoreData(int write_slow)
 
   WriteData(lstAddr, PointingData[i_point].lst, NIOS_QUEUE);
 
-  WriteData(xMagAddr, (unsigned int)(ACSData.mag_x), NIOS_QUEUE);
-  WriteData(yMagAddr, (unsigned int)(ACSData.mag_y), NIOS_QUEUE);
-  WriteData(zMagAddr, (unsigned int)(ACSData.mag_z), NIOS_QUEUE);
+  WriteData(xMagAddr, ACSData.mag_x, NIOS_QUEUE);
+  WriteData(yMagAddr, ACSData.mag_y, NIOS_QUEUE);
+  WriteData(zMagAddr, ACSData.mag_z, NIOS_QUEUE);
   WriteData(azMagAddr,
       (unsigned int)((PointingData[i_point].mag_az +
                       CommandData.mag_az_trim) * DEG2I), NIOS_QUEUE);
